+++ data/hts/entd/filtered.py
@@ -8,19 +8,22 @@ This stage filters out ENTD observations which live or work outside of
 
 def configure(context):
     context.stage("data.hts.entd.cleaned")
+    context.stage("data.spatial.codes")
 
 def execute(context):
+    df_codes = context.stage("data.spatial.codes")
     df_households, df_persons, df_trips = context.stage("data.hts.entd.cleaned")
 
     # Filter for non-residents
-    f = df_persons["departement_id"].isin(hts.FILTER_DEPARTEMENTS)
+    requested_departments = df_codes["departement_id"].unique()
+    f = df_persons["departement_id"].astype(str).isin(requested_departments)
     df_persons = df_persons[f]
 
     # Filter for people going outside of the area (because they have NaN distances)
     remove_ids = set()
 
     remove_ids |= set(df_trips[
-        ~df_trips["origin_departement_id"].isin(hts.FILTER_DEPARTEMENTS) | ~df_trips["destination_departement_id"].isin(hts.FILTER_DEPARTEMENTS)
+        ~df_trips["origin_departement_id"].astype(str).isin(requested_departments) | ~df_trips["destination_departement_id"].astype(str).isin(requested_departments)
     ]["person_id"].unique())
 
     df_persons = df_persons[~df_persons["person_id"].isin(remove_ids)]