+++ data/spatial/centroid_distances.py
@@ -1,12 +1,10 @@
 import pandas as pd
 
 def configure(context):
-    context.stage("data.spatial.zones")
+    context.stage("data.spatial.municipalities")
 
 def execute(context):
-    df = context.stage("data.spatial.zones")
-    df = df[df["zone_level"] == "commune"]
-
+    df = context.stage("data.spatial.municipalities")
     records = []
 
     with context.progress(total = len(df)**2, label = "Calculating centroid distances ...") as progress: