+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 **1.0.0-dev**
 
+- Make pipeline configurable for other departments and regions, add documentation
+- BC: Make use of INSEE zone summary data (`codes_2017`)
+- Add configuration parameters to filter for departments and regions
+- Fixed destinations that have coordinates outside of their municipality
 - Make error message for runtime dependencies more verbose
 - Switch default instructions to Anaconda